+++ rewrite-java-core/src/test/java/com/azure/recipes/core/v2/AddTryCatchToMethodCallUnitTest.java
@@ -15,23 +15,22 @@ import static org.openrewrite.java.Assertions.java;
  * Tests simple method renaming with declarative recipe.
  * @author Annabelle Mittendorf Smith
  */
-public class CatchUncheckedExceptionsTest implements RewriteTest {
+public class AddTryCatchToMethodCallUnitTest implements RewriteTest {
     /**
      * This method defines recipes used for testing.
      * @param spec stores settings for testing environment.
      */
     @Override
     public void defaults(RecipeSpec spec) {
-        spec.recipes(new CatchUncheckedExceptions("UserClass myMethod(..)", true,"e.printStackTrace();"));
+        spec.recipes(new AddTryCatchToMethodCallRecipe("UserClass myMethod(..)", true,"e.printStackTrace();"));
     }
 
     @Test
     void test_addTryCatch_voidMethod() {
-        @Language("java") String before ="import java.io.IOException;\n" +
-                "\n" +
+        @Language("java") String before =
                 "public class UserClass {\n" +
                 "    \n" +
-                "    private void myMethod() {\n" + // throws IOException
+                "    private void myMethod() {\n" +
                 "        int a = 1 + 1;\n" +
                 "    }\n" +
                 "    \n" +
@@ -46,7 +45,7 @@ public class CatchUncheckedExceptionsTest implements RewriteTest {
                 "\n" +
                 "public class UserClass {\n" +
                 "    \n" +
-                "    private void myMethod() {\n" + // throws IOException
+                "    private void myMethod() {\n" +
                 "        int a = 1 + 1;\n" +
                 "    }\n" +
                 "    \n" +
@@ -68,11 +67,10 @@ public class CatchUncheckedExceptionsTest implements RewriteTest {
 
     @Test
     void test_addTryCatch_methodInVarDeclaration() {
-        @Language("java") String before ="import java.io.IOException;\n" +
-                "\n" +
+        @Language("java") String before =
                 "public class UserClass {\n" +
                 "    \n" +
-                "    private int myMethod() {\n" + // throws IOException
+                "    private int myMethod() {\n" +
                 "       return 2;\n" +
                 "    }\n" +
                 "    \n" +
@@ -85,7 +83,7 @@ public class CatchUncheckedExceptionsTest implements RewriteTest {
                 "\n" +
                 "public class UserClass {\n" +
                 "    \n" +
-                "    private int myMethod() {\n" + // throws IOException
+                "    private int myMethod() {\n" +
                 "       return 2;\n" +
                 "    }\n" +
                 "    \n" +
@@ -106,11 +104,10 @@ public class CatchUncheckedExceptionsTest implements RewriteTest {
 
     @Test
     void test_addTryCatch_methodInAssignment() {
-        @Language("java") String before ="import java.io.IOException;\n" +
-                "\n" +
+        @Language("java") String before =
                 "public class UserClass {\n" +
                 "    \n" +
-                "    private int myMethod() {\n" + // throws IOException
+                "    private int myMethod() {\n" +
                 "       return 2;\n" +
                 "    }\n" +
                 "    \n" +
@@ -146,14 +143,13 @@ public class CatchUncheckedExceptionsTest implements RewriteTest {
     }
 
     @Test
-    void test_addTryCatch_methodIsMemberVariable() {
-        @Language("java") String before ="import java.io.IOException;\n" +
-                "\n" +
+    void test_addTryCatch_methodIsFromInstance() {
+        @Language("java") String before =
                 "public class UserClass {\n" +
                 "    public UserClass(){}\n" +
                 "    String s = \"Hello\";\n" +
                 "    \n" +
-                "    public String myMethod() {\n" + // throws IOException
+                "    public String myMethod() {\n" +
                 "       return s;\n" +
                 "    }\n" +
                 "}\n" +
@@ -161,7 +157,7 @@ public class CatchUncheckedExceptionsTest implements RewriteTest {
                 "class UserClass2 {" +
                 "    \n" +
                 "    public void myMethod2() {\n" +
-                "        UserClass c = new UserClass();\n" + // throws IOException
+                "        UserClass c = new UserClass();\n" +
                 "        String s2 = c.myMethod();\n" +
                 "    }\n" +
                 "}\n";
@@ -172,7 +168,7 @@ public class CatchUncheckedExceptionsTest implements RewriteTest {
                 "    public UserClass(){}\n" +
                 "    String s = \"Hello\";\n" +
                 "    \n" +
-                "    public String myMethod() {\n" + // throws IOException
+                "    public String myMethod() {\n" +
                 "       return s;\n" +
                 "    }\n" +
                 "}\n" +
@@ -180,7 +176,7 @@ public class CatchUncheckedExceptionsTest implements RewriteTest {
                 "class UserClass2 {" +
                 "    \n" +
                 "    public void myMethod2() {\n" +
-                "        UserClass c = new UserClass();\n" + // throws IOException
+                "        UserClass c = new UserClass();\n" +
                 "        String s2 = null;\n" +
                 "        try {\n" +
                 "            s2 = c.myMethod();\n" +
@@ -189,6 +185,59 @@ public class CatchUncheckedExceptionsTest implements RewriteTest {
                 "        }\n" +
                 "    }\n" +
                 "}\n";
+
+        rewriteRun(
+                java(before,after)
+        );
+    }
+
+    @Test
+    void test_addTryCatch_methodIsInNestedCall() {
+        @Language("java") String before =
+                "public class UserClass {\n" +
+                "    public UserClass(){}\n" +
+                "    String s = \"Hello\";\n" +
+                "    \n" +
+                "    public String myMethod() {\n" +
+                "       return s;\n" +
+                "    }\n" +
+                "}\n" +
+                "\n" +
+                "class UserClass2 {\n" +
+                "    UserClass c = new UserClass();\n" +
+                "}\n" +
+                "class UserClass3 {\n" +
+                "    UserClass2 c2 = new UserClass2();\n" +
+                "    public void myMethod3() {\n" +
+                "        String s = c2.c.myMethod();\n" +
+                "    }\n" +
+                "}";
+
+        @Language("java") String after = "import java.io.IOException;\n" +
+                "\n" +
+                "public class UserClass {\n" +
+                "    public UserClass(){}\n" +
+                "    String s = \"Hello\";\n" +
+                "    \n" +
+                "    public String myMethod() {\n" +
+                "       return s;\n" +
+                "    }\n" +
+                "}\n" +
+                "\n" +
+                "class UserClass2 {\n" +
+                "    UserClass c = new UserClass();\n" +
+                "}\n" +
+                "class UserClass3 {\n" +
+                "    UserClass2 c2 = new UserClass2();\n" +
+                "    public void myMethod3() {\n" +
+                "        String s = null;\n" +
+                "        try {\n" +
+                "            s = c2.c.myMethod();\n" +
+                "        } catch (IOException e) {\n" +
+                "            e.printStackTrace();\n" +
+                "        }\n" +
+                "    }\n" +
+                "}";
 
         rewriteRun(
                 java(before,after)
